requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
Babel==2.15.0
base58==2.1.1
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
differential-privacy==0.1.0
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
dparse==0.6.4b0
email_validator==2.2.0
entrypoints==0.4
etils==1.5.2
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
featuretools==1.31.0
filelock==3.12.4
Flask==2.2.5
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
holidays==0.53
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
intel-openmp==2021.4.0; platform_machine == 'x86_64'
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.3
isort==5.10.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-packaging==0.7.12
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab==3.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.16.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
langchain==0.2.9
langchain-community==0.2.7
langchain-core==0.2.21
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.92
language_data==1.2.0
libclang==18.1.1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0; platform_machine == 'x86_64'
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbclassic==0.5.6
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.15
opencv-python==4.8.0.76
opt-einsum==3.3.0
optax==0.2.1
optree==0.12.1
orbax-checkpoint==0.4.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
Pillow==9.4.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
python-multipart==0.0.7
pytz==2024.1
pywin32==306; sys_platform == 'win32'
pywinpty==2.0.13; sys_platform == 'win32'
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
safety==3.2.4
safety-schemas==0.0.2
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.46.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==0.80.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0; sys_platform != 'darwin' or platform_machine != 'arm64'
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==1.2.3
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
validators==0.33.0
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
woodwork==0.31.0
wrapt==1.14.1
yarl==1.9.4
zipp==3.19.2
